Grassfire works now

+++ src/grassfire.cpp
@@ -13,7 +13,7 @@ using std::max;
 
 //Forward declaration required due to CFFI's requirement to have unmangled symbols
 extern "C" {
-    int grassfire(uint32_t* dst, const uint8_t* mask, size_t width, size_t height);
+    int grassfire(uint32_t* dst, const uint8_t* mask, uint32_t width, uint32_t height);
 }
 
 /**
@@ -21,14 +21,14 @@ extern "C" {
  * Takes a destination counter array (must be zero-initialized)
  * and a binary mask array (checked for != 0).
  */
-int grassfire(uint32_t* dst, const uint8_t* mask, size_t width, size_t height) {
+int grassfire(uint32_t* dst, const uint8_t* mask, uint32_t width, uint32_t height) {
     // 1st pass
-    for (int x = 0; x < width; ++x) {
-        for (int y = 0; y < height; ++y) {
+    for (uint32_t x = 0; x < width; ++x) {
+        for (uint32_t y = 0; y < height; ++y) {
             if (IMG_XY(mask, x, y) != 0) { // Pixel in contour
                 // Get neighbors
-                uint32_t north = (y == 0) ? 0 : IMG_XY(dst, x, y - 1);
-                uint32_t west = (x == 0) ? 0 : IMG_XY(dst, x - 1, y);
+                int north = (y == 0) ? 0 : IMG_XY(dst, x, y - 1);
+                int west = (x == 0) ? 0 : IMG_XY(dst, x - 1, y);
                 // Set value
                 IMG_XY(dst, x, y) = 1 + min(north, west);
             }
